bitloader/requirements.txt
celery==3.1.23
python-gflags==2.0
pymongo==3.2.2
redis==2.10.5
pyzmq==15.2.0
gevent==1.1.1
dnspython==1.12.0
tornado==4.3
base58==0.2.2
geoip2==2.3.0
sqlalchemy==1.0.12
pycrypto==2.6.1
MySQL-python==1.2.5
IPy==0.81